organized doc-specific functions
